requirements.txt
tensorflow==2.20.0
opencv-python
matplotlib
numpy
streamlit
pillow